Added the VertexBufferComponent and Vertex class. SceneManager now handles the drawing of any entity with the VertexBuffer component, including setting the right position and rotation.

+++ include/Quaternion.h
@@ -59,7 +59,7 @@ namespace luck { namespace core {
             Vector3<T> vn(v);
             vn.normalize();
 
-            sinAngle = sin(angle);
+            sinAngle = sinf(angle);
 
             x = (vn.x * sinAngle);
             y = (vn.y * sinAngle);
